Add PackStore: append-only pack file storage for build cache (#230)
* cache: FUSE virtual filesystem local tier (packed bodies, no tiny files)

Turn the GOCACHEPROG local cache into a virtual filesystem so the build
cache stops being thousands of tiny files (a data file + a .meta sidecar
per entry). The GET response hands the compiler a DiskPath it opens
itself, and nothing requires that path to be a real file: back it with a
read-only FUSE mount and the kernel materializes each body on demand from
an append-only pack.

- PackStore (pack.go): append-only, content-addressed pack files with
  self-describing records; index rebuilt by scanning at startup (no
  separate index file); torn final records detected and dropped
  (crash-safe); outputID dedup; rotation + size-bounded reset. Portable
  and unit-tested without FUSE.
- FuseCache (fusecache.go, linux/darwin): mounts read-only FUSE; Get
  returns <mnt>/<outputID>; reads served via PackStore.ReadAt. Mounts via
  go-fuse DirectMount (mount(2) as root) with fusermount fallback for
  non-root CI. Stable inode per content; positive entries cached, negative
  lookups never cached (the store only grows).
- LocalStore interface (store.go) + NewLocalStore selector: graceful
  fallback to the loose-file LocalCache when FUSE is unavailable
  (Windows / no /dev/fuse / no permission).
- Single-owner flock on the cache dir: a second daemon (e.g. a nested
  go-toolchain run in tests sharing XDG_CACHE_HOME) falls back to the
  loose cache instead of unmounting the live mount.
- docs/CACHE.md: architecture, protocol, pack format, and request flows
  with mermaid diagrams; README + CLAUDE.md updated.


* cache: persist content-dedup aliases so warm builds stay local

The FUSE local tier deduped identical content (same outputID) by updating only
the in-memory actionID->location map and writing nothing to disk. On the next
build the index is rebuilt by scanning pack records, so every deduped action --
thousands of them, since every empty output shares sha256("") -- was missing and
fell through to the (slow) network tier. That roughly doubled the warm build
(52s vs 25s loose locally; 41s in CI, over the 30s second-build gate).

Persist the dedup as a header-only alias record (actionID -> outputID, second
magic) so it survives a restart. Warm builds now serve entirely from the local
pack: 0 web hits, ~20s.

Also serve FUSE reads zero-copy via ReadResultFd and raise max_read/readahead to
1 MiB, cutting per-read round-trips for the compiler's mmap'd archives.


* cache: fix darwin cross-compile + address review hardening

The darwin matrix target failed to build: syscall.MNT_FORCE is undefined on
darwin (it lives in golang.org/x/sys/unix). Latent until now — action-test was
skipped while the second-build gate failed; with the gate passing it ran and
exposed it. Use unix.MNT_FORCE.

Review hardening (PR #230):
- pack.go ReadAt: reject negative offsets (would underflow past the header).
- pack.go ReadAll: bound body length before the int64->int make().
- pack.go scanPack: require alias dataLen==0 (else stop, don't desync), and
  truncate a stranded torn tail so a crash can't inflate disk/reset accounting.
- pack.go append: write header and body as separate WriteAts, dropping the
  second full-body copy for large archives.
- fusecache.go flock: treat only EWOULDBLOCK as "busy"; surface real errors.
- fusecache.go Close: Wait() whenever an unmount succeeded (incl. clean path),
  and never Wait() when all unmount attempts failed (would block forever).
- comment fixes: newLocalStore -> NewLocalStore.


---------

+++ src/cmd/deps_test.go
@@ -311,7 +311,7 @@ func TestDepChecker_checkDep_CacheExpired(t *testing.T) {
 	err = db.QueryRow(`SELECT checked_at FROM deps WHERE path = ? AND version = ?`,
 		"github.com/spf13/cobra", "v1.10.2").Scan(&checkedAt)
 	require.Nil(t, err)
-	assert.NotEqual(t, 0, checkedAt)
+	assert.NotEqual(t, int64(0), checkedAt)
 }
 
 func TestDepChecker_run_Canceled(t *testing.T) {
